chore: rename configuration to constants
