+++ packages/middleware-flexible-checksums/src/index.ts
@@ -1,4 +1,4 @@
-export * from "./configuration";
+export * from "./constants";
 export * from "./flexibleChecksumsMiddleware";
 export * from "./getFlexibleChecksumsPlugin";
 export * from "./resolveFlexibleChecksumsConfig";
