+++ packages/middleware-flexible-checksums/src/selectChecksumAlgorithmFunction.spec.ts
@@ -1,7 +1,7 @@
 import { AwsCrc32 } from "@aws-crypto/crc32";
 import { AwsCrc32c } from "@aws-crypto/crc32c";
 
-import { ChecksumAlgorithm } from "./configuration";
+import { ChecksumAlgorithm } from "./constants";
 import { selectChecksumAlgorithmFunction } from "./selectChecksumAlgorithmFunction";
 
 describe(selectChecksumAlgorithmFunction.name, () => {
