+++ packages/middleware-flexible-checksums/src/types.ts
@@ -1,4 +1,4 @@
-import { ChecksumAlgorithm } from "./configuration";
+import { ChecksumAlgorithm } from "./constants";
 
 /**
  * List of algorithms supported by client.
