+++ packages/middleware-flexible-checksums/src/validateChecksumFromResponse.ts
@@ -1,7 +1,7 @@
 import { HttpResponse } from "@aws-sdk/protocol-http";
 import { HashConstructor } from "@aws-sdk/types";
 
-import { ChecksumAlgorithm } from "./configuration";
+import { ChecksumAlgorithm } from "./constants";
 import { getChecksum } from "./getChecksum";
 import { getChecksumAlgorithmListForResponse } from "./getChecksumAlgorithmListForResponse";
 import { getChecksumLocationName } from "./getChecksumLocationName";
